+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 2.0.0-SNAPSHOT
+## 2.0.0
 
 - Removes `Codec` api
 - Rename `ValueOption` to `ToOption` in `IResult<T>` interface